+++ e2e-tests/content-strings/quote-titles.js
@@ -4,7 +4,7 @@ export const QUOTE_TITLES = {
   [FIELD_IDS.CONTRACT_VALUE]: 'Contract value',
   [FIELD_IDS.MAX_AMOUNT_OWED]: 'Maximum buyer will owe at any point',
   [`${[FIELD_IDS.QUOTE.INSURED_FOR]}_SINGLE_POLICY`]: "You'll be insured for",
-  [`${[FIELD_IDS.QUOTE.INSURED_FOR]}_MULTI_POLICY`]: "Maximum you'll be insured for",
+  [`${[FIELD_IDS.QUOTE.INSURED_FOR]}_MULTIPLE_POLICY`]: "Maximum you'll be insured for",
   [FIELD_IDS.PERCENTAGE_OF_COVER]: 'Percentage of cover',
   [FIELD_IDS.QUOTE.PREMIUM_RATE_PERCENTAGE]: 'Premium rate percentage',
   [FIELD_IDS.QUOTE.ESTIMATED_COST]: 'Estimated cost to you',