+++ src/ui/server/constants/field-ids/shared/index.ts
@@ -1,7 +1,7 @@
 const SHARED = {
   POLICY_TYPE: 'policyType',
   SINGLE_POLICY_TYPE: 'singlePolicyType',
-  MULTI_POLICY_TYPE: 'multiPolicyType',
+  MULTIPLE_POLICY_TYPE: 'multiplePolicyType',
 };
 
 export default SHARED;