+++ src/ui/server/controllers/insurance/policy-and-export/type-of-policy/index.ts
@@ -106,7 +106,7 @@ export const post = async (req: Request, res: Response) => {
 
     return res.redirect(ROUTES.PROBLEM_WITH_SERVICE);
   } catch (err) {
-    console.error('Error updating application', { err });
+    console.error('Error updating application - policy and exports - type of policy', { err });
 
     return res.redirect(ROUTES.PROBLEM_WITH_SERVICE);
   }