+++ src/ui/server/helpers/data-content-mappings/map-policy-length.ts
@@ -2,7 +2,7 @@ import { isSinglePolicyType, isMultiPolicyType } from '../policy-type';
 import mapMonthString from './map-month-string';
 import { FIELD_IDS } from '../../constants';
 
-const { POLICY_TYPE, POLICY_LENGTH, SINGLE_POLICY_LENGTH, MULTI_POLICY_LENGTH } = FIELD_IDS;
+const { POLICY_TYPE, POLICY_LENGTH, SINGLE_POLICY_LENGTH, MULTIPLE_POLICY_LENGTH } = FIELD_IDS;
 
 /**
  * mapPolicyLength
@@ -21,7 +21,7 @@ const mapPolicyLength = (data: object) => {
 
   if (isMultiPolicyType(data[POLICY_TYPE])) {
     mapped = {
-      [MULTI_POLICY_LENGTH]: mapMonthString(data[POLICY_LENGTH]),
+      [MULTIPLE_POLICY_LENGTH]: mapMonthString(data[POLICY_LENGTH]),
     };
   }