+++ src/ui/types/company-details.d.ts
@@ -1,5 +1,3 @@
-import { SummaryListItemData } from './summary-list';
-
 interface CompanyDetails {
   companyRegistration?: string;
   companyName?: string;
@@ -8,13 +6,9 @@ interface CompanyDetails {
   companySIC?: string;
 }
 
-interface CompanyDetailsFieldGroups {
-  COMPANY_DETAILS: Array<SummaryListItemData>;
-}
-
 interface SicCode {
   id: string;
   sicCode: string;
 }
 
-export { CompanyDetails, CompanyDetailsFieldGroups, SicCode };
+export { CompanyDetails, SicCode };